--- a/gatsby-config.js
+++ b/gatsby-config.js
@@ -18,7 +18,14 @@ module.exports = {
     },
     'gatsby-plugin-sharp',
     'gatsby-transformer-sharp',
-    'gatsby-source-shopify-experimental',
+    {
+      resolve: 'gatsby-source-shopify-experimental',
+      options: {
+        apiKey: process.env.SHOPIFY_ADMIN_API_KEY,
+        password: process.env.SHOPIFY_ADMIN_PASSWORD,
+        storeUrl: process.env.SHOPIFY_STORE_URL
+      }
+    },
     'gatsby-plugin-react-helmet',
     'gatsby-plugin-emotion',
     {