--- a/gatsby-config.js
+++ b/gatsby-config.js
@@ -23,7 +23,7 @@ module.exports = {
       options: {
         apiKey: process.env.SHOPIFY_ADMIN_API_KEY,
         password: process.env.SHOPIFY_ADMIN_PASSWORD,
-        storeUrl: process.env.SHOPIFY_STORE_URL
+        storeUrl: 'gatsby-swag.myshopify.com'
       }
     },
     'gatsby-plugin-react-helmet',