Note that this will currently require an .env file with these keys in order to test:

```
SHOPIFY_STORE_URL=gatsby-swag.myshopify.com
SHOPIFY_ACCESS_TOKEN=
SHOPIFY_ADMIN_API_KEY=
SHOPIFY_ADMIN_PASSWORD=
```

I'll publish another release that accepts these values as plugin options.

## Benchmarking

Tracking build time comparisons here as I go. Might delete later IDK.

### Old plugin

```
success open and validate gatsby-configs - 0.056s
success load plugins - 0.454s
success onPreInit - 0.024s
success initialize cache - 0.005s
success copy gatsby files - 0.019s
success onPreBootstrap - 0.011s
success createSchemaCustomization - 0.004s

gatsby-source-shopify/gatsby-swag starting to fetch data from Shopify

gatsby-source-shopify/gatsby-swag fetched and processed Collection nodes: 489.448ms

gatsby-source-shopify/gatsby-swag fetched and processed ShopPolicy nodes: 479.597ms

gatsby-source-shopify/gatsby-swag fetched and processed Blog nodes: 488.670ms

gatsby-source-shopify/gatsby-swag fetched and processed Article nodes: 587.472ms

gatsby-source-shopify/gatsby-swag fetched and processed Page nodes: 600.137ms
success Downloading remote files - 3.478s - 61/61 17.54/s

gatsby-source-shopify/gatsby-swag fetched and processed Product nodes: 6019.998ms

gatsby-source-shopify/gatsby-swag finished fetching data from Shopify: 6017.212ms
success Checking for changed pages - 0.001s
success source and transform nodes - 6.065s
success building schema - 0.311s
info Total nodes: 366, SitePage nodes: 34 (use --verbose for breakdown)
success createPages - 0.035s
success Checking for changed pages - 0.001s
success createPagesStatefully - 0.037s
success update schema - 0.019s
success write out redirect data - 0.001s
success Build manifest and related icons - 0.083s
success onPostBootstrap - 0.086s
info bootstrap finished - 8.516s
success onPreExtractQueries - 0.001s
success extract queries from components - 0.236s
success write out requires - 0.003s
success run static queries - 3.598s - 3/3 0.83/s
success run page queries - 0.012s - 3/3 240.81/s
success Generating image thumbnails - 176.699s - 1156/1156 6.54/s
```

### New plugin

```
success open and validate gatsby-configs - 0.054s
success load plugins - 0.435s
success onPreInit - 0.022s
success initialize cache - 0.006s
success copy gatsby files - 0.018s
success onPreBootstrap - 0.009s
success createSchemaCustomization - 0.005s
Checked for operations in progress: 569.551ms
Initiated bulk operation query: 376.147ms
Completed bulk operation: 5619.528ms
Created nodes from bulk operation: 2378.194ms
createProductsOperation: 9807.659ms
success Checking for changed pages - 0.001s
success source and transform nodes - 9.839s
success building schema - 0.599s
info Total nodes: 494, SitePage nodes: 36 (use --verbose for breakdown)
success createPages - 0.039s
success Checking for changed pages - 0.000s
success createPagesStatefully - 0.043s
success update schema - 0.017s
success write out redirect data - 0.001s
success Build manifest and related icons - 0.086s
success onPostBootstrap - 0.089s
info bootstrap finished - 12.490s
success onPreExtractQueries - 0.001s
success extract queries from components - 0.301s
success write out requires - 0.007s
success run static queries - 3.510s - 3/3 0.85/s
success run page queries - 0.026s - 3/3 113.65/s
success Generating image thumbnails - 194.428s - 1196/1196 6.15/s
```