@sslotsky 2021-01-26
> Note that this will require an .env file with these keys:

We should probably publish a new version of the plugin that accepts those values as plugin options, but either way, we'll need all the values