@sslotsky 2021-01-29
Check out the benchmarking results. What's odd to me here is that we're not fetching as many data types, but we wind up with more nodes.

It also takes a little bit longer to source nodes in the tests I've run so far. I think this is partly from waiting for the bulk operation to complete. I wouldn't be surprised if this sort of thing slows down builds for smaller Shopify sites by a little bit, while hopefully increasing both speed and reliability for larger sites by a lot.