@KyleAMathews 2021-01-29
>It also takes a little bit longer to source nodes in the tests I've run so far. I think this is partly from waiting for the bulk operation to complete.

What are the numbers?

How often do we run the bulk operation? Presumably if we already have one cached locally we just do a delta fetch?