@sslotsky 2021-01-29
> > It also takes a little bit longer to source nodes in the tests I've run so far. I think this is partly from waiting for the bulk operation to complete.
> 
> What are the numbers?
> 
> How often do we run the bulk operation? Presumably if we already have one cached locally we just do a delta fetch?

Check out the PR description for lots of numbers, including these

```
success source and transform nodes - 6.065s
```
```
success source and transform nodes - 9.839s
```

And yes, if we have one cached, we'll do the delta fetch. That should be an improvement for sure.