@sslotsky 2021-01-29
I added some breakdowns for the timing of `sourceAllNodes` that can also be seen in the results that I pasted in the description:

```
Checked for operations in progress: 569.551ms
Initiated bulk operation query: 376.147ms
Completed bulk operation: 5619.528ms
Created nodes from bulk operation: 2378.194ms
createProductsOperation: 9807.659ms
```

`createProductsOperation` encapsulates the previous four lines. Polling the bulk operation talks ~5.6 seconds. Once that completes it takes ~2.4 seconds to create the nodes.