@sslotsky 2021-02-01
So we have a dilemma here regarding the secrets. The new plugin uses the admin API, and the keys to that API do include some write access for customer data. So we probably do not want to add them to the repo. 

I'm not sure if there's a way to create a new private app with strictly read-only access for the admin API. Might have to enable private app development in order to create a new one?