@sslotsky 2021-02-01
After looking over the branch build there are a couple issues that need investigating. 

1. A couple additional products show up
2. Some products show up in a different order
3. Some products feature different images
4. Products with more than one image only seem to process the first image correctly